Minor bug fixes in (collision) distance field (#1392)
* Update collision_world_distance_field.h


deleted  void generateEnvironmentDistanceField(bool redo = true); 
No source exists.

* Update voxel_grid.h

fixed worldtogrid and gridtoworld world input vectors to doubles as required.

* Update collision_robot_hybrid.h

replaced hardcoded defaults

* Update collision_world_hybrid.h

replaced hardcoded defaults with variables from robot_distance_field.

* apply clang-format

+++ moveit_core/distance_field/include/moveit/distance_field/voxel_grid.h
@@ -243,7 +243,7 @@ public:
    * @return True, as there is no check that the integer locations are valid
    */
   void gridToWorld(int x, int y, int z, double& world_x, double& world_y, double& world_z) const;
-  void gridToWorld(const Eigen::Vector3i& grid, Eigen::Vector3i& world) const;
+  void gridToWorld(const Eigen::Vector3i& grid, Eigen::Vector3d& world) const;
 
   /**
    * \brief Converts from a world location to a set of integer
@@ -262,7 +262,7 @@ public:
    * that pass a validity check; otherwise False.
    */
   bool worldToGrid(double world_x, double world_y, double world_z, int& x, int& y, int& z) const;
-  bool worldToGrid(const Eigen::Vector3i& world, Eigen::Vector3i& grid) const;
+  bool worldToGrid(const Eigen::Vector3d& world, Eigen::Vector3i& grid) const;
 
   /**
    * \brief Checks if the given cell in integer coordinates is within the voxel grid
@@ -553,7 +553,7 @@ inline void VoxelGrid<T>::gridToWorld(int x, int y, int z, double& world_x, doub
 }
 
 template <typename T>
-inline void VoxelGrid<T>::gridToWorld(const Eigen::Vector3i& grid, Eigen::Vector3i& world) const
+inline void VoxelGrid<T>::gridToWorld(const Eigen::Vector3i& grid, Eigen::Vector3d& world) const
 {
   world.x() = getLocationFromCell(DIM_X, grid.x());
   world.y() = getLocationFromCell(DIM_Y, grid.y());
@@ -570,7 +570,7 @@ inline bool VoxelGrid<T>::worldToGrid(double world_x, double world_y, double wor
 }
 
 template <typename T>
-inline bool VoxelGrid<T>::worldToGrid(const Eigen::Vector3i& world, Eigen::Vector3i& grid) const
+inline bool VoxelGrid<T>::worldToGrid(const Eigen::Vector3d& world, Eigen::Vector3i& grid) const
 {
   grid.x() = getCellFromLocation(DIM_X, world.x());
   grid.y() = getCellFromLocation(DIM_Y, world.y());
